requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
